+++ frontend/src/page/body_fat.rs
@@ -445,13 +445,18 @@ pub fn view(model: &Model, data_model: &data::Model) -> Node<Msg> {
     if data_model.body_fat.is_empty() && data_model.loading_body_fat {
         common::view_loading()
     } else {
+        let dates = data_model.body_fat.values().map(|w| w.date);
+        let body_fat_interval = common::Interval {
+            first: dates.clone().min().unwrap_or_default(),
+            last: dates.max().unwrap_or_default(),
+        };
         div![
             view_body_fat_dialog(
                 &model.dialog,
                 model.loading,
                 data_model.session.as_ref().unwrap().sex
             ),
-            common::view_interval_buttons(&model.interval, Msg::ChangeInterval),
+            common::view_interval_buttons(&model.interval, &body_fat_interval, Msg::ChangeInterval),
             view_chart(model, data_model),
             view_calendar(data_model, &model.interval),
             view_table(model, data_model),